+++ hat/examples/dft/src/main/java/dft/Main.java
@@ -111,8 +111,8 @@ public class Main {
             for (int k = 0; k < size; k++) {
                 float angle = -2 * HATMath.PI * ((k * idx) % size) / size;
                 Complex complexInput = input.complex(k);
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (complexInput.real() * cReal) - (complexInput.imag() * cImag);
                 sumImag += (complexInput.real() * cImag) + (complexInput.imag() * cReal);
             }
@@ -137,8 +137,8 @@ public class Main {
             float sumImag = 0.0f;
             for (int k = 0; k < size; k++) {
                 float angle = -2 * HATMath.PI * ((idx * k) % size) / size;
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (inReal.array(k) * cReal) - (inImag.array(k) * cImag);
                 sumImag += (inReal.array(k) * cImag) + (inImag.array(k) * cReal);
             }
@@ -164,8 +164,8 @@ public class Main {
             for (int j = 0; j < size; j++) {
                 float angle = -2 * HATMath.PI * ((j * k) % size) / size;
                 Complex complexInput = input.complex(j);
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (complexInput.real() * cReal) - (complexInput.imag() * cImag);
                 sumImag += (complexInput.real() * cImag) + (complexInput.imag() * cReal);
             }
@@ -182,8 +182,8 @@ public class Main {
             for (int k = 0; k < size; k++) {
                 float angle = -2 * HATMath.PI * ((idx * k) % size) / size;
                 Complex complexInput = input.complex(k);
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (complexInput.real() * cReal) - (complexInput.imag() * cImag);
                 sumImag += (complexInput.real() * cImag) + (complexInput.imag() * cReal);
             }
