+++ hat/tests/src/main/java/hat/test/TestDFT.java
@@ -80,8 +80,8 @@ public class TestDFT {
             for (int k = 0; k < size; k++) {
                 float angle = -2 * HATMath.PI * ((k * idx) % size) / size;
                 Complex complexInput = input.complex(k);
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (complexInput.real() * cReal) - (complexInput.imag() * cImag);
                 sumImag += (complexInput.real() * cImag) + (complexInput.imag() * cReal);
             }
@@ -102,8 +102,8 @@ public class TestDFT {
             for (int j = 0; j < size; j++) {
                 float angle = -2 * HATMath.PI * ((j * k) % size) / size;
                 Complex complexInput = input.complex(j);
-                float cReal = HATMath.cos(angle);
-                float cImag = HATMath.sin(angle);
+                float cReal = HATMath.cosf(angle);
+                float cImag = HATMath.sinf(angle);
                 sumReal += (complexInput.real() * cReal) - (complexInput.imag() * cImag);
                 sumImag += (complexInput.real() * cImag) + (complexInput.imag() * cReal);
             }
